+++ clients/spot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 31.0.0 - 2026-05-12
+
+### Added (1)
+
+#### WebSocket Streams
+
+- `blockTrade()` (`<symbol>@blockTrade` stream)
+
 ## 30.0.1 - 2026-05-07
 
 ### Changed (1)